pyproject.toml
[project]
name = "logscope"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "github-copilot-sdk>=0.2.2",
    "presidio-analyzer>=2.2.362",
    "presidio-anonymizer>=2.2.362",
    "spacy>=3.4.4",
    "en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
logscope = "logscope.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[tool.hatch.build.targets.wheel]
artifacts = ["src/logscope/_meta.py"]

[tool.hatch.build.targets.sdist]
artifacts = ["src/logscope/_meta.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/logscope"]

[tool.coverage.report]
fail_under = 90

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
